Rename bap.py to run.py
