+++ plugins/bap/utils/__init__.py
@@ -1,4 +1,4 @@
 """Commonly used utilities."""
 
-__all__ = ('sexpr', 'bap_comment', 'bap', 'ida', 'abstract_ida_plugins',
+__all__ = ('sexpr', 'bap_comment', 'run', 'ida', 'abstract_ida_plugins',
            'config')
